requirements.txt
ibm-watsonx-ai==1.3.37
langchain==0.3.27
langchain-ibm==0.3.19
langchain-community==0.3.31
sentence-transformers==2.2.2
chromadb==1.1.1
pypdf==6.1.1
SQLAlchemy==2.0.44
gradio==4.36.1
python-dotenv==1.1.1
